+++ srv/v3/query.go
@@ -32,7 +32,7 @@ func Query(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	// Evaluate the query
-	queryResult := make(map[int]struct{})
+	queryResult := make(map[uint64]struct{})
 	if err := db.EvalQuery(qJson, dbcol, &queryResult); err != nil {
 		http.Error(w, fmt.Sprint(err), 400)
 		return
@@ -79,7 +79,7 @@ func Count(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, fmt.Sprintf("Collection '%s' does not exist.", col), 400)
 		return
 	}
-	queryResult := make(map[int]struct{})
+	queryResult := make(map[uint64]struct{})
 	if err := db.EvalQuery(qJson, dbcol, &queryResult); err != nil {
 		http.Error(w, fmt.Sprint(err), 400)
 		return
